constants cleanup
