+++ src/main/java/com/tobqol/rooms/maiden/commons/MaidenTable.java
@@ -23,7 +23,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.tobqol.rooms.maiden.commons.util;
+package com.tobqol.rooms.maiden.commons;
 
 import com.google.common.collect.ImmutableTable;
 import com.google.common.collect.Table;
